+++ lihil/vendors.py
@@ -3,6 +3,6 @@ from starlette.datastructures import FormData as FormData
 from starlette.datastructures import Headers as Headers
 from starlette.datastructures import QueryParams as QueryParams
 from starlette.datastructures import UploadFile as UploadFile
-from starlette.requests import Request as Request
+from starlette.requests import Request as Request, cookie_parser as cookie_parser
 from starlette.responses import Response as Response
 from starlette.types import Lifespan as Lifespan